Added docstring to metalog.py changed variable names to reflect python objects
Replaced _list suffix with _dict to make the variable names jive with python style not R.

Picked the google docstring style and added it for metalog.__init__ class. This is the style I will use to do all the other methods and classes in the package.

Also created and successfully tested setup.py script which will allow us to generate a wheel out of the package (first step to pip).