Added L2 Regularization to Metalog OLS Fit Method
Added an L2 Regularization term to the Metalog linear regression fit method, essentially turning the Metalog linear regression fit method into a Ridge Regression. This feature prevents overfitting and allows for the user to pick the highest term limit when using the returned metalog object with no degradation to performance.